Fix C++ warnings
Check for exactly 1 before using the singular noun, rather than check for greater than 1, just in case somehow the variable being evaluated ends up as 0

In addition, should this code ever be refactored in such a way that the check to 0 is changed, this will ensure the end result does not change. In addition, it is naturally immediately apparent to the developers what is going on, potentially enhancing maintainability.

Additionally, make bitwise operations more obvious in their intentions.

Fix unused parameters in functions by using [[maybe_unused]]

+++ src/CalculatorUnitTests/StandardViewModelUnitTests.cpp
@@ -728,7 +728,7 @@ namespace CalculatorUnitTests
             viewModel->DisplayValue = L"1001";
             viewModel->OnMemoryButtonPressed();
             viewModel->OnMemoryButtonPressed();
-            VERIFY_ARE_EQUAL((int)viewModel->MemorizedNumbers->Size, 2);
+            VERIFY_ARE_EQUAL(viewModel->MemorizedNumbers->Size, 2);
         }
 
         // When memory list is empty and M+ is pressed
@@ -900,7 +900,7 @@ namespace CalculatorUnitTests
             {
                 viewModel->OnMemoryButtonPressed();
             }
-            VERIFY_ARE_EQUAL((int)viewModel->MemorizedNumbers->Size, 100);
+            VERIFY_ARE_EQUAL(viewModel->MemorizedNumbers->Size, 100);
         }
 
         // When memory slot is pressed verify if the display value is updated correctly
